+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Second version after feedback
 - Fix: changed native input to shadcn input in Service Card.
 - Update: add not established to service request client preferences, timeframes and time
 - Fix: Include only esitmated price into service card.
+- Feautre: Default value inputs & editor